requirements.txt
# Dev environment
pip
autopep8

# App
flask
requests
python-dotenv
